fix(hogql): accept full Int64 range for interval string count
ClickHouse stores intervals as Int64 (DataTypeInterval is DataTypeNumberBase<Int64>)
and handles the whole Int64 range correctly; beyond it the value silently overflows.
The combined-string interval form `INTERVAL '<count> <unit>'` was capped at int32 by
the cpp parser's std::stoi, while the equivalent `INTERVAL <count> <unit>` and
`toIntervalX(<count>)` forms already accepted the full range. Align all three forms
by converting the string count with std::stoll (cpp) / i64 (rust), so e.g.
`INTERVAL '3000000000 day'` now parses on both backends and only counts past Int64
max are rejected (matching std::stoll's out_of_range).

- common/hogql_parser/parser_json.cpp: std::stoi -> std::stoll (int -> int64_t)
- rust emit_interval_combined_string: parse the count as i64; error text matches
  cpp's stoll what() ("Unknown error: stoll: ...")
- bumps hogql-parser 1.3.72 -> 1.3.73 and hogql-parser-rs 1.3.102 -> 1.3.103

Verified by building both parsers locally and diffing rust vs cpp across the string,
expr+unit, and direct-call forms over the int32/Int64 boundaries. Runtime-confirmed
the Int64 range against the ClickHouse cluster.

+++ common/hogql_parser/parser_json.cpp
@@ -1826,7 +1826,8 @@ class HogQLParseTreeJSONConverter : public HogQLParserBaseVisitor {
         throw NotImplementedError(("Unsupported interval count: " + count_str).c_str());
       }
     }
-    int countInt = std::stoi(count_str);
+    // ClickHouse stores intervals as Int64, so accept the full Int64 range (stoll), not just int32 (stoi).
+    int64_t countInt = std::stoll(count_str);
 
     std::string name;
     if (unit_str == "second" || unit_str == "seconds") {
